Add Parser, Storage, Ui, Commands classes
