+++ src/data/duke.txt
@@ -1,0 +1,3 @@
+todo me
+todo you
+event me /from 27-05-2001 1800/to 27-05-2001 1900
