+++ src/main/java/Commands.java
@@ -1,0 +1,5 @@
+package main.java;
+
+public class Commands {
+    enum COMMANDS {BYE, LIST, TODO, DEADLINE, EVENT, MARK, UNMARK, DELETE, BY, FROM, TO, SORT, FIND, UNKNOWN}
+}
