--- src/main/java/Load.java
@@ -1,5 +1,0 @@
-package main.java;
-
-public class Load {
-
-}
